subsys/settings: change processing to const char
The settings module processes the variable name by splitting it up in
a set of variables. This PR removes the splitting up and keeps the
variable name as one string.

The PR is an alternative to #16609

The PR also introduces the possibility to register handler including a
separator, or to register a handler for each variable.

The PR also introduces the ability to load a subtree from flash or even
to load a single item.

Signed-off-by: Laczen JMS <laczenjms@gmail.com>